# colexec: improve single-group hash agg performance