--- a/pkg/sql/colexec/hash_aggregator.eg.go
+++ b/pkg/sql/colexec/hash_aggregator.eg.go
@@ -13,6 +13,7 @@ import (
 	"bytes"
 	"fmt"
 	"math"
+	"sort"
 
 	"github.com/cockroachdb/cockroach/pkg/col/coldata"
 	"github.com/cockroachdb/cockroach/pkg/col/coldataext"
@@ -30,40 +31,37 @@ func (op *hashAggregator) populateSels(b coldata.Batch, hashBuffer []uint64) {
 	// they all are of zero length here (see the comment for op.scratch.sels
 	// for context).
 	op.scratch.hashCodeForSelsSlot = op.scratch.hashCodeForSelsSlot[:0]
+	sort.Slice(hashBuffer, func(i, j int) bool { return hashBuffer[i] < hashBuffer[j] })
 	if batchSelection := b.Selection(); batchSelection != nil {
-		for selIdx, hashCode := range hashBuffer {
-			selsSlot := -1
-			for slot, hash := range op.scratch.hashCodeForSelsSlot {
-				if hash == hashCode {
-					// We have already seen a tuple with the same hashCode
-					// previously, so we will append into the same sels slot.
-					selsSlot = slot
-					break
-				}
-			}
-			if selsSlot < 0 {
-				// This is the first tuple in hashBuffer with this hashCode, so we
-				// will add this tuple to the next available sels slot.
-				selsSlot = len(op.scratch.hashCodeForSelsSlot)
+		if len(hashBuffer) == 0 {
+			return
+		}
+		// The first hashcode is always new to us.
+		op.scratch.hashCodeForSelsSlot = append(op.scratch.hashCodeForSelsSlot, hashBuffer[0])
+		op.scratch.sels[0] = append(op.scratch.sels[0], batchSelection[0])
+		selsSlot := 0
+		for selIdx := 1; selIdx < len(hashBuffer); selIdx++ {
+			hashCode := hashBuffer[selIdx]
+			if op.scratch.hashCodeForSelsSlot[selsSlot] != hashCode {
+				// We've advanced to a new value in hashBuffer. Make a new slot.
+				selsSlot++
 				op.scratch.hashCodeForSelsSlot = append(op.scratch.hashCodeForSelsSlot, hashCode)
 			}
 			op.scratch.sels[selsSlot] = append(op.scratch.sels[selsSlot], batchSelection[selIdx])
 		}
 	} else {
-		for selIdx, hashCode := range hashBuffer {
-			selsSlot := -1
-			for slot, hash := range op.scratch.hashCodeForSelsSlot {
-				if hash == hashCode {
-					// We have already seen a tuple with the same hashCode
-					// previously, so we will append into the same sels slot.
-					selsSlot = slot
-					break
-				}
-			}
-			if selsSlot < 0 {
-				// This is the first tuple in hashBuffer with this hashCode, so we
-				// will add this tuple to the next available sels slot.
-				selsSlot = len(op.scratch.hashCodeForSelsSlot)
+		if len(hashBuffer) == 0 {
+			return
+		}
+		// The first hashcode is always new to us.
+		op.scratch.hashCodeForSelsSlot = append(op.scratch.hashCodeForSelsSlot, hashBuffer[0])
+		op.scratch.sels[0] = append(op.scratch.sels[0], 0)
+		selsSlot := 0
+		for selIdx := 1; selIdx < len(hashBuffer); selIdx++ {
+			hashCode := hashBuffer[selIdx]
+			if op.scratch.hashCodeForSelsSlot[selsSlot] != hashCode {
+				// We've advanced to a new value in hashBuffer. Make a new slot.
+				selsSlot++
 				op.scratch.hashCodeForSelsSlot = append(op.scratch.hashCodeForSelsSlot, hashCode)
 			}
 			op.scratch.sels[selsSlot] = append(op.scratch.sels[selsSlot], selIdx)