I noticed that most of the time in the hash agg is currently spent doing
an N^2 search for equal items in the hash buffer, so I added a sort
phase to the hash buffer to avoid that.

This gives a 30% improvement to the single group case and 10% worse
performance for larger group case. Not sure if it's worth it.

```
name                                                                         old time/op    new time/op    delta
Aggregator/SUM/hash/int/groupSize=1/hasNulls=false/numInputBatches=64-24        126ms ± 3%      94ms ± 2%  -25.04%  (p=0.000 n=10+9)
Aggregator/SUM/hash/int/groupSize=1/hasNulls=true/numInputBatches=64-24         126ms ± 3%      94ms ± 4%  -25.32%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=false/numInputBatches=64-24       67.5ms ± 2%    56.2ms ± 4%  -16.77%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=true/numInputBatches=64-24        66.6ms ± 2%    54.4ms ± 3%  -18.40%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=false/numInputBatches=64-24     7.39ms ± 1%    8.34ms ± 1%  +12.88%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=true/numInputBatches=64-24      6.95ms ± 1%    7.84ms ± 0%  +12.74%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=false/numInputBatches=64-24    7.27ms ± 1%    7.91ms ± 1%   +8.79%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=true/numInputBatches=64-24     6.86ms ± 1%    7.44ms ± 1%   +8.38%  (p=0.000 n=10+10)

name                                                                         old speed      new speed      delta
Aggregator/SUM/hash/int/groupSize=1/hasNulls=false/numInputBatches=64-24     4.16MB/s ± 3%  5.53MB/s ± 3%  +32.90%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1/hasNulls=true/numInputBatches=64-24      4.17MB/s ± 3%  5.59MB/s ± 4%  +33.98%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=false/numInputBatches=64-24     7.77MB/s ± 2%  9.33MB/s ± 4%  +20.19%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=true/numInputBatches=64-24      7.87MB/s ± 2%  9.64MB/s ± 3%  +22.56%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=false/numInputBatches=64-24   71.0MB/s ± 1%  62.9MB/s ± 1%  -11.41%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=true/numInputBatches=64-24    75.4MB/s ± 1%  66.9MB/s ± 0%  -11.30%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=false/numInputBatches=64-24  72.1MB/s ± 1%  66.3MB/s ± 1%   -8.09%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=true/numInputBatches=64-24   76.4MB/s ± 1%  70.5MB/s ± 1%   -7.73%  (p=0.000 n=10+10)

name                                                                         old alloc/op   new alloc/op   delta
Aggregator/SUM/hash/int/groupSize=1/hasNulls=false/numInputBatches=64-24       31.3MB ± 0%    31.2MB ± 0%   -0.29%  (p=0.000 n=10+9)
Aggregator/SUM/hash/int/groupSize=1/hasNulls=true/numInputBatches=64-24        31.2MB ± 0%    31.1MB ± 0%   -0.32%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=false/numInputBatches=64-24       16.4MB ± 0%    16.3MB ± 0%   -0.06%  (p=0.004 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=true/numInputBatches=64-24        16.2MB ± 0%    16.2MB ± 0%   -0.04%  (p=0.001 n=10+8)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=false/numInputBatches=64-24      116kB ± 0%     120kB ± 0%   +3.53%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=true/numInputBatches=64-24       116kB ± 0%     120kB ± 0%   +3.54%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=false/numInputBatches=64-24    82.6kB ± 0%    86.7kB ± 0%   +4.96%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=true/numInputBatches=64-24     82.6kB ± 0%    86.7kB ± 0%   +4.96%  (p=0.000 n=10+10)

name                                                                         old allocs/op  new allocs/op  delta
Aggregator/SUM/hash/int/groupSize=1/hasNulls=false/numInputBatches=64-24         204k ± 0%      204k ± 0%   +0.05%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1/hasNulls=true/numInputBatches=64-24          191k ± 0%      191k ± 0%     ~     (p=0.052 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=false/numInputBatches=64-24         118k ± 0%      119k ± 0%   +0.14%  (p=0.004 n=10+10)
Aggregator/SUM/hash/int/groupSize=2/hasNulls=true/numInputBatches=64-24          115k ± 0%      115k ± 0%   +0.17%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=false/numInputBatches=64-24      1.17k ± 0%     1.30k ± 0%  +10.90%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=512/hasNulls=true/numInputBatches=64-24       1.17k ± 0%     1.30k ± 0%  +10.94%  (p=0.000 n=8+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=false/numInputBatches=64-24       910 ± 0%      1038 ± 0%  +14.07%  (p=0.000 n=10+10)
Aggregator/SUM/hash/int/groupSize=1024/hasNulls=true/numInputBatches=64-24        910 ± 0%      1038 ± 0%  +14.07%  (p=0.000 n=10+10)
```