@yuzefovich 2020-07-05
I think it’s a good idea, just an incomplete implementation - we simply need to additionally switch the indices in the selection vector of the buffered batch when sorting hashBuffer, and it’ll probably work (if the selection is not yet set, we'll need to populate it with “identity” selection vector.