@yuzefovich 2020-07-07
I wrote up a commit that does the swapping of the selection vector of the batch in #51039, and the benchmarks are now expectedly not as clear-cut. I still think it's probably a beneficial change given the absolute speeds in big group sizes are still quite high. One interesting thing to look at is the speed comparison for the case when the selection vector is already present on the batch.