@jordanlewis 2021-10-08
I'm fairly sure this is closeable... it's ancient, and I'm definitely going to pick it up, and presumably the code has changed a ton since... any reason to keep these ideas around @yuzefovich?